src/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
requests>=2.31.0
reportlab>=4.0.0
python-barcode[images]>=0.15.1
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
django-filter>=23.5
